[python] refactor: Refactor processor - 추상화하여 코드 베이스를 통일함. - process -> __call__로 entry 변경 및 로직개선
Refs: #3
