+++ pretrain_language_model.py
@@ -115,7 +115,7 @@ def main():
 def _mp_fn(index):
     # For xla_spawn (TPUs)
     main()
-    
+
 
 if __name__ == "__main__":
     main()
