+++ xla_spawn.py
@@ -43,7 +43,9 @@ def parse_args():
     )
 
     # Optional arguments for the launch helper
-    parser.add_argument("--num_cores", type=int, default=1, help="Number of TPU cores to use (1 or 8).")
+    parser.add_argument(
+        "--num_cores", type=int, default=1, help="Number of TPU cores to use (1 or 8)."
+    )
 
     # positional
     parser.add_argument(
@@ -73,7 +75,11 @@ def main():
     mod = importlib.import_module(mod_name)
 
     # Patch sys.argv
-    sys.argv = [args.training_script] + args.training_script_args + ["--tpu_num_cores", str(args.num_cores)]
+    sys.argv = (
+        [args.training_script]
+        + args.training_script_args
+        + ["--tpu_num_cores", str(args.num_cores)]
+    )
 
     xmp.spawn(mod._mp_fn, args=(), nprocs=args.num_cores)
